+++ sources/packages/backend/apps/queue-consumers/src/processors/schedulers/institution-integration/ier12-integration/_tests_/e2e/models/data-inputs/programs-inputs.ts
@@ -16,7 +16,9 @@ export const PROGRAM_UNDERGRADUATE_CERTIFICATE_WITHOUT_INSTITUTION_PROGRAM_CODE:
 export const PROGRAM_GRADUATE_DIPLOMA_WITH_INSTITUTION_PROGRAM_CODE: IER12Program =
   {
     name: "Some Program With Description Too Long",
-    description: "Some program with description too long",
+    // Added new line character expecting to be replaced during
+    // file processing.
+    description: "Some program with\n description too long",
     credentialType: "graduateDiploma",
     fieldOfStudyCode: 1,
     cipCode: "12.1234",
