+++ sources/packages/backend/libs/utilities/src/index.ts
@@ -13,3 +13,4 @@ export * from "./parse-json";
 export * from "./integration-utils";
 export * from "./math-utils";
 export * from "./specialized-string-builder";
+export * from "./string-utils";
