+++ sources/packages/backend/libs/utilities/src/integration-utils.ts
@@ -1,4 +1,3 @@
-const REPLACE_LINE_BREAK_REGEX = /\r?\n|\r/g;
 /**
  * Express the completion years data present on educational programs as
  * an amount of years.
@@ -22,21 +21,4 @@ export function getTotalYearsOfStudy(completionYears: string): number {
     default:
       return 1;
   }
-}
-
-/**
- * Replace the line breaks in the given text.
- * @param data data to replace the line break.
- * @param options replace options:
- * - `replaceText`: text to replace the line break.
- * @returns line break replaced string.
- */
-export function replaceLineBreaks(
-  data?: string,
-  options?: { replaceText?: string },
-) {
-  if (!data) {
-    return data;
-  }
-  return data.replace(REPLACE_LINE_BREAK_REGEX, options?.replaceText ?? "");
 }
